4 new hostile NPCs (spawn on dungeon floors 1 & 2 in Plains biome)
* Dungeon generator now fills map with default tile before generating. This fixes occasional bug with grass tiles in dungeon
* Lake generator  from & to params adjustment
* Abandoned gas stations now have a 30% chance to have a vending machine instead of spawning it always

+++ MadSand-core/src/hitonoriol/madsand/world/worldgen/LakePreset.java
@@ -4,5 +4,5 @@ public class LakePreset {
 	public int lakeTile;
 	public int lakeRadius;
 	public float lakeModifier;
-	public double lakeFrom, lakeTo;
+	public float lakeFrom, lakeTo;
 }
